+++ source/Matrix.h
@@ -4,8 +4,8 @@
 
 struct Matrix
 {
-	std::vector<double> di;
-	std::vector<double> ggl;
+	std::vector<float> di;
+	std::vector<float> ggl;
 	std::vector<int> ig;
 	std::vector<int> jg;
 };
